feat: コーヒー評価の星を 0.5 刻み（ハーフスター）対応（iOS）
StarRatingView を Double + ハーフスター（star.leadinghalf.filled）表示に。
編集は星を左右2分割した透明タップ領域で 0.5/1.0 を判定（StarTapCell）。
VoiceOver は 0.5 刻みで増減、値は「3.5星」表記。

- CoffeeEditor/Detail/List/CafeDetail の rating を Double に追随
- CoffeeFirestoreMapper.swift: rating を Double 書き込み、読みは NSNumber
  経由フォールバックで旧 Int 保存データも互換
- PreviewSamples: rating を Double（半端値含む）に

検証: xcodebuild（framework ビルド込み）BUILD SUCCEEDED

+++ iosApp/iosApp/Features/CoffeeEditor/CoffeeEditorViewModelBridge.swift
@@ -74,8 +74,8 @@ final class CoffeeEditorViewModelBridge {
         kotlin.onVisitedOnChanged(date: date)
     }
 
-    func onRatingChanged(rating: Int) {
-        kotlin.onRatingChanged(rating: Int32(rating))
+    func onRatingChanged(rating: Double) {
+        kotlin.onRatingChanged(rating: rating)
     }
 
     func onNotesChanged(_ text: String) {
